+++ src/main/java/modelo/dao/MetodoPagoDAO.java
@@ -7,8 +7,8 @@ public interface MetodoPagoDAO {
     public void create(MetodoPago mp) throws Exception;
     public List<MetodoPago> read() throws Exception;
     public void update(MetodoPago mp) throws Exception;
-    public void delete(long id) throws Exception;
+    public void delete(int id) throws Exception;
     
     public List<String> obtenerNombresMetodosPago() throws Exception;
-    public long obtenerIdPorNombre(String nombre) throws Exception;
+    public int obtenerIdPorNombre(String nombre) throws Exception;
 }
